% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAnalysis.R
\name{doAnalysis}
\alias{doAnalysis}
\title{perform an analysis of a sample object}
\usage{
doAnalysis(
  sample = doSample(autoShow = FALSE),
  evidence = braw.def$evidence,
  autoShow = FALSE
)
}
\value{
analysis object
}
\description{
perform an analysis of a sample object
}
\examples{
analysis<-doAnalysis(sample=doSample(),evidence=makeEvidence(),autoShow=braw.env$autoShow)#' make a multiple samples
}
