% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doExplore.R
\name{doExplore}
\alias{doExplore}
\title{make multiple samples whilst varying a parameter}
\usage{
doExplore(
  nsims = 10,
  exploreResult = NA,
  explore = braw.def$explore,
  hypothesis = braw.def$hypothesis,
  design = braw.def$design,
  evidence = braw.def$evidence,
  metaAnalysis = braw.def$metaAnalysis,
  doingNull = FALSE,
  doingMetaAnalysis = FALSE,
  autoShow = FALSE,
  showType = "rs"
)
}
\arguments{
\item{exploreType}{"rIV","Heteroscedasticity","rIV2","rIVIV2","rIVIV2DV" \cr
"p(R+)","Lambda" \cr
"n","Method","Usage","WithinCorr","ClusterRad","SampleSD" \cr
 "Dependence","Outliers","NonResponse","IVRange","IVRangeC","IVRangeE","DVRange" \cr
 "Cheating","CheatingAmount" \cr
 "Alpha","Transform","InteractionOn" \cr
 "Power","Keep","Repeats" \cr}
}
\value{
exploreResult object
}
\description{
make multiple samples whilst varying a parameter
}
\examples{
exploreResult<-doExplore(nsims=10,exploreResult=NULL,explore=braw.def$explore,
                             doingNull=FALSE,autoShow=FALSE,showType="Basic")
}
\seealso{
showExplore()

reportExplore()
}
