% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMetaAnalysis.R
\name{doMetaAnalysis}
\alias{doMetaAnalysis}
\title{do a meta-analysis}
\usage{
doMetaAnalysis(
  metaSingle = braw.res$metaSingle,
  metaAnalysis = braw.def$metaAnalysis,
  keepStudies = FALSE,
  shortHand = TRUE,
  hypothesis = braw.def$hypothesis,
  design = braw.def$design,
  evidence = braw.def$evidence
)
}
\value{
metaResult object
}
\description{
do a meta-analysis
}
\examples{
doMetaAnalysis<-function(metaSingle=NULL,metaAnalysis=makeMetaAnalysis(),
                         keepStudies=FALSE,shortHand=TRUE,
                         hypothesis=braw.def$hypothesis,design=braw.def$design,evidence=braw.def$evidence)
}
