% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMetaAnalysis.R
\name{doMetaMultiple}
\alias{doMetaMultiple}
\title{do multiple meta-analyses}
\usage{
doMetaMultiple(
  nsims = 100,
  metaMultiple = braw.res$metaMultiple,
  metaAnalysis = braw.def$metaAnalysis,
  shortHand = TRUE,
  hypothesis = braw.def$hypothesis,
  design = braw.def$design,
  evidence = braw.def$evidence
)
}
\value{
metaResult object
}
\description{
do multiple meta-analyses
}
\examples{
doMetaMultiple<-function(nsims=100,metaMultiple=braw.res$metaMultiple,metaAnalysis=braw.def$metaAnalysis,
                         shortHand=TRUE,
                         hypothesis=braw.def$hypothesis,design=braw.def$design,evidence=braw.def$evidence)
}
