% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMultiple.R
\name{doMultiple}
\alias{doMultiple}
\title{make multiple samples with analysis}
\usage{
doMultiple(
  nsims = 10,
  multipleResult = NA,
  hypothesis = braw.def$hypothesis,
  design = braw.def$design,
  evidence = braw.def$evidence,
  doingNull = FALSE,
  inSteps = FALSE,
  autoShow = braw.env$autoShow,
  showType = "Basic",
  onlyReplication = FALSE
)
}
\value{
multipleResult object
}
\description{
make multiple samples with analysis
}
\examples{
multipleResult<-doMultiple(nsims=100,multipleResult=NULL,hypothesis=makeHypothesis(),design=makeDesign(),evidence=makeEvidence(),
                             doingNull=FALSE,inSteps=FALSE,autoShow=braw.env$autoShow,showType="Basic")
}
\seealso{
showMultiple() and reportMultiple())
}
