% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPossible.R
\name{doPossible}
\alias{doPossible}
\title{likelihood theory}
\usage{
doPossible(possible = NULL, possibleResult = NULL)
}
\value{
possibleResult object
}
\description{
likelihood theory
}
\examples{
doPossible<-function(possible=makePossible(),possibleResult=NULL)
}
\seealso{
showPossible()
}
