% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAnalysis.R
\name{doSingle}
\alias{doSingle}
\title{makes a sample and analyses it}
\usage{
doSingle(
  hypothesis = braw.def$hypothesis,
  design = braw.def$design,
  evidence = braw.def$evidence,
  onlyReplication = FALSE,
  autoShow = braw.env$autoShow
)
}
\value{
result object
}
\description{
makes a sample and analyses it
}
\examples{
analysis<-doSingle(hypothesis=makeHypothesis(),design=makeDesign(),evidence=makeEvidence(),autoShow=braw.env$autoShow)#' make a multiple samples
}
