% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMultiple.R
\name{extractMultiple}
\alias{extractMultiple}
\title{ex6ract multiple samples}
\usage{
extractMultiple(r1, n = NULL)
}
\value{
multipleResult object
}
\description{
ex6ract multiple samples
}
\examples{
newResult<-extractMultiple(mres1,n)
}
