% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSemModel.R
\name{fit_sem_model}
\alias{fit_sem_model}
\title{fit a SEM model to existing data}
\usage{
fit_sem_model(pathmodel, model_data, fixedCoeffs = NULL, doinglavaan = TRUE)
}
\value{
sample object
}
\description{
fit a SEM model to existing data
}
\examples{
fit_sem_model<-function(pathmodel,model_data,fixedCoeffs=NULL)
}
