% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAnalysis.R
\name{generalAnalysis}
\alias{generalAnalysis}
\title{analyse a sample}
\usage{
generalAnalysis(
  allData,
  AnalysisTerms,
  withins = FALSE,
  ssqType = "Type3",
  caseOrder = "AsStated"
)
}
\value{
result object
}
\description{
analyse a sample
}
\examples{
generalAnalysis<-function(allData,InteractionOn,withins=FALSE,ssqType="Type3",caseOrder="AsStated")
}
