% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSystem.R
\name{getDesign}
\alias{getDesign}
\title{make a specific design}
\usage{
getDesign(name, design = braw.def$design)
}
\arguments{
\item{name}{"Psych"}
}
\value{
hypothesis object
}
\description{
make a specific design
}
\examples{
design<-getDesign(name,design=makeDesign())
}
