% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSystem.R
\name{getHypothesis}
\alias{getHypothesis}
\title{make a specific hypothesis}
\usage{
getHypothesis(name, hypothesis = braw.def$hypothesis)
}
\arguments{
\item{name}{"Psych","3"}
}
\value{
hypothesis object
}
\description{
make a specific hypothesis
}
\examples{
hypothesis<-getHypothesis(name,hypothesis=makeHypothesis())
}
