% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVariable.R
\name{getVariable}
\alias{getVariable}
\title{make a specific variable}
\usage{
getVariable(name)
}
\arguments{
\item{name}{"Psych","Treatment","Treatment?","IQ","Diligence","Perfectionism", \cr
"Happiness","Grade","RiskTaking","Interesting","Coffee?","Smoker?", \cr
"RiskTaker?","Musician?","StudySubject","BirthOrder"}
}
\value{
a Variable structure
}
\description{
make a specific variable
}
\examples{
variable<-getVariable(name)
}
\seealso{
showVariable(variable=makeVariable("test"))
}
