% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSystem.R
\name{getWorld}
\alias{getWorld}
\title{make a specific hypothesis}
\usage{
getWorld(name, rp = 0.3, result = braw.res$result)
}
\arguments{
\item{name}{"Psych","PsychF"}
}
\value{
world object
}
\description{
make a specific hypothesis
}
\examples{
world<-getWorld(name)
}
