% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnalysis.R
\name{makeAnalysis}
\alias{makeAnalysis}
\title{perform an analysis of a sample object}
\usage{
makeAnalysis(
  sample = makeSample(),
  evidence = makeEvidence(),
  autoShow = FALSE
)
}
\value{
analysis object
}
\description{
perform an analysis of a sample object
}
\examples{
analysis<-makeAnalysis(sample=makeSample(),evidence=makeEvidence(),autoShow=FALSE)#' make a multiple samples
}
