% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{makeEffect}
\alias{makeEffect}
\title{make an effect object}
\usage{
makeEffect(
  rIV = 0,
  rIV2 = 0,
  rIVIV2 = 0,
  rIVIV2DV = 0,
  rSD = 0,
  rM1 = NULL,
  rM2 = NULL,
  Heteroscedasticity = 0,
  ResidDistr = "normal",
  world = makeWorld(FALSE)
)
}
\arguments{
\item{ResidDistr}{"normal","skewed","uniform","cauchy","t(3)"}
}
\value{
an effect object
}
\description{
make an effect object
}
\examples{
makeEffect(rIV=0.3,rIV2=0,rIVIV2=0,rIVIV2DV=0,rSD=0,Heteroscedasticity=0,
           ResidDistr="normal",world=makeWorld()
)
}
