% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doExplore.R
\name{doExplore}
\alias{doExplore}
\title{make multiple samples whilst varying a parameter}
\usage{
doExplore(
  nsims = 10,
  exploreResult = NULL,
  exploreType = "n",
  exploreNPoints = 13,
  min_n = 10,
  max_n = 250,
  max_r = 0.9,
  max_anom = 1,
  xlog = FALSE,
  xabs = FALSE,
  mx_log = FALSE,
  hypothesis = makeHypothesis(),
  design = makeDesign(),
  evidence = makeEvidence(),
  doingNull = FALSE,
  autoShow = FALSE,
  showType = "r"
)
}
\arguments{
\item{exploreType}{"rIV","Heteroscedasticity","rIV2","rIVIV2","rIVIV2DV" \cr
"n","Method","Usage","WithinCorr","ClusterRad","SampleGamma" \cr
 "Dependence","Outliers","IVRange","DVRange" \cr
 "Cheating","CheatingAmount" \cr
 "Alpha","Transform" \cr
 "Power","SigOnly","Repeats" \cr}
}
\value{
exploreResult object
}
\description{
make multiple samples whilst varying a parameter
}
\examples{
exploreResult<-doExplore(nsims=10,exploreResult=NULL,exploreType="n",exploreNPoints=13,
                             min_n=10,max_n=250,max_r=0.9,max_anom=1,
                             xlog=FALSE,xabs=FALSE,mx_log=FALSE,
                             hypothesis=makeHypothesis(),design=makeDesign(),evidence=makeEvidence(),
                             doingNull=FALSE,autoShow=braw.env$autoShow,showType="Basic")
}
\seealso{
showExplore()

reportExplore()
}
