% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePossible.R
\name{makePossible}
\alias{makePossible}
\title{likelihood theory}
\usage{
makePossible(
  typePossible = "Samples",
  possibleResult = NULL,
  UseSource = "world",
  targetSample = 0.3,
  UsePrior = "none",
  prior = getWorld("Psych"),
  targetPopulation = 0.3,
  hypothesis = makeHypothesis(),
  design = makeDesign(),
  simSlice = 0.1,
  correction = TRUE,
  appendSim = FALSE,
  possibleLength = "10"
)
}
\arguments{
\item{typePossible}{"Samples","Populations"}
}
\value{
possibleResult object
}
\description{
likelihood theory
}
\examples{
makePossible<-function(typePossible="Samples",
possibleResult=NULL,
UseSource="world",targetSample=0.3,
UsePrior="none",prior=getWorld("Psych"),targetPopulation=0.3,
hypothesis=makeHypothesis(),design=makeDesign(),
simSlice=0.1,correction=TRUE,
appendSim=FALSE,possibleLength="10")
}
\seealso{
showPossible()
}
