% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{makeReplication}
\alias{makeReplication}
\title{make a replication object}
\usage{
makeReplication(
  ReplicationOn = FALSE,
  ReplRepeats = 1,
  ReplKeep = "last",
  ReplPowerOn = TRUE,
  ReplPower = 0.8,
  ReplTails = 2,
  ReplSigOnly = FALSE,
  ReplBudgetType = "Fixed",
  ReplBudget = 1000,
  ReplPowerPrior = "None",
  ReplVarAlpha = FALSE,
  ReplAlphaChange = 2
)
}
\arguments{
\item{ReplKeep}{"last", "cautious", "largeN", "smallP", "median"}

\item{ReplBudgetType}{"Fixed", "Budget"}

\item{ReplPowerPrior}{"None", "World", "Prior"}
}
\value{
a replication object
}
\description{
make a replication object
}
\examples{
makeReplication(ReplicationOn=FALSE,ReplRepeats=1,ReplKeep="last",
                ReplPowerOn=TRUE,ReplPower=0.8,ReplTails=2,
                ReplSigOnly="No",
                ReplBudgetType="Fixed",ReplBudget=1000,
                ReplPowerPrior="None",
                ReplVarAlpha=FALSE,ReplAlphaChange=2
                )
}
