% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSample.R
\name{makeSample}
\alias{makeSample}
\title{make a simulated sample}
\usage{
makeSample(
  hypothesis = makeHypothesis(),
  design = makeDesign(),
  autoShow = FALSE
)
}
\value{
a sample object
}
\description{
make a simulated sample
}
\examples{
sample<-makeSample(hypothesis=makeHypothesis(),design=makeDesign(),autoShow=braw.env$autoShow)
}
\seealso{
showSample() reportSample()
}
