% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{makeSampling}
\alias{makeSampling}
\title{make a sampling method object}
\usage{
makeSampling(type = "Random")
}
\arguments{
\item{type}{"Random","Limited,"Stratified","Cluster","Snowball","Convenience"}
}
\value{
an effect object
}
\description{
"Random"
   purely random sample from whole range
"Stratified"
  sampled at specific intervals
  
"Cluster"
  a number of clusters, 
  each cluster having a particular radius within the population
  within each cluster a number of members
  random sampling within each cluster
  
"Snowball"
  a number of clusters,
  each cluster having a particular radius within the population
  within each cluster a number of members
  from each cluster member a chain of contacts
  each contact having a particular radius from their predecessor
  
"Convenience"
  like Snowball but more clusters and shorter chains of contacts
}
\examples{
makeSampling(type="Random")
}
