% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSEM.R
\name{makeSEMPath}
\alias{makeSEMPath}
\title{make path for a SEM model}
\usage{
makeSEMPath(
  data = NULL,
  stages = NULL,
  depth = 1,
  only_ivs = NULL,
  only_dvs = NULL,
  within_stages = 0,
  add = NULL,
  remove = NULL
)
}
\value{
path object
}
\description{
make path for a SEM model
}
\examples{
makeSEMPath<-function(data,stages=NULL)
}
