% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVariable.R
\name{makeVariable}
\alias{makeVariable}
\title{make a variable}
\usage{
makeVariable(
  name,
  type = "Interval",
  mu = 0,
  sd = 1,
  skew = 0,
  kurtosis = 3,
  nlevs = 7,
  iqr = 3,
  median = NULL,
  discrete = "discrete",
  ordProportions = NA,
  ncats = 2,
  cases = "C1,C2",
  proportions = "1,1",
  source = "discrete",
  deploy = "Between",
  targetDeploys = "",
  process = "sim"
)
}
\arguments{
\item{type}{"Interval","Ordinal","Categorical"}
}
\description{
make a variable
}
\examples{
makeVariable(name,type="Interval",
             mu=0,sd=1,skew=0,kurtosis=3,
             nlevs=7,iqr=3,median=NULL,discrete="discrete",ordProportions=NA,
             ncats=2,cases="C1,C2",proportions="1,1",source="discrete",
             deploy="Between",targetDeploys="",process="sim"
)
}
