% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{makeWorld}
\alias{makeWorld}
\title{make a world
 an object that specifies the distribution of effect sizes}
\usage{
makeWorld(
  worldOn = TRUE,
  populationPDF = "Uniform",
  populationRZ = "r",
  populationPDFk = 0,
  populationPDFmu = 0,
  populationNullp = 0,
  populationPDFsample = FALSE,
  populationSamplemn = 0,
  populationSamplesd = 0,
  populationSamplebias = FALSE,
  sigOnly = FALSE,
  worldAbs = FALSE
)
}
\arguments{
\item{populationPDF}{"Single","Double","Uniform","Gauss","Exp"}

\item{populationRZ}{"r","z"}
}
\value{
a world object
}
\description{
make a world
 an object that specifies the distribution of effect sizes
}
\examples{
makeWorld<-function(worldOn=FALSE,populationPDF="Single",populationRZ="r",
                    populationPDFk=0.2,populationNullp=0,worldAbs=FALSE
)
}
\seealso{
showWorld(world=makeWorld())
}
