% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMultiple.R
\name{mergeMultiple}
\alias{mergeMultiple}
\title{merge multiple samples}
\usage{
mergeMultiple(r1, r2)
}
\value{
multipleResult object
}
\description{
merge multiple samples
}
\examples{
multipleResult<-mergeMultiple(mres1,mres2)
}
