% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicPlot.R
\name{nullPlot}
\alias{nullPlot}
\title{this is the start of any figure - which may then contain several graphs}
\usage{
nullPlot()
}
\value{
a plot object
}
\description{
this is the start of any figure - which may then contain several graphs
}
\examples{
g<-nullPlot()
}
