% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSemModel.R
\name{plotSEMModel}
\alias{plotSEMModel}
\title{plot a fitted SEM model}
\usage{
plotSEMModel(pathmodel)
}
\value{
sample object
}
\description{
plot a fitted SEM model
}
\examples{
plotSEMModel<-function(pathmodel)
}
