% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRead.R
\name{prepareSample}
\alias{prepareSample}
\title{prepare an existing sample}
\usage{
prepareSample(raw_data, doOrdinals = FALSE, maxOrdinal = 9, header = c())
}
\value{
sample object
}
\description{
prepare an existing sample
}
\examples{
prepareSample<-function(raw_data, doOrdinals=FALSE, maxOrdinal=9, header=c())
}
