% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSEM.R
\name{readDataSEM}
\alias{readDataSEM}
\title{read data for a SEM model}
\usage{
readDataSEM(filename)
}
\value{
sample object
}
\description{
read data for a SEM model
}
\examples{
readDataSEM<-function(filename)
}
