% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRead.R
\name{readSample}
\alias{readSample}
\title{prepare an existing sample}
\usage{
readSample(data, DV = NULL, IV = NULL, IV2 = NULL)
}
\value{
sample object
}
\description{
prepare an existing sample
}
\examples{
readSample<-function(data,DV,IV,IV2=NULL)
}
