% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDescription.R
\name{reportDescription}
\alias{reportDescription}
\title{report a sample effect-size analysis of a simulated sample}
\usage{
reportDescription(analysis = braw.res$result, plain = FALSE)
}
\value{
ggplot2 object - and printed
}
\description{
report a sample effect-size analysis of a simulated sample
}
\examples{
reportDescription(analysis=doAnalysis())
}
