% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSystem.R
\name{reportDesign}
\alias{reportDesign}
\title{report a design}
\usage{
reportDesign(design = braw.def$design, plain = FALSE)
}
\value{
ggplot2 object - and printed
}
\description{
report a design
}
\examples{
reportDesign(design=braw.def$design)
}
