% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportExpected.R
\name{reportExpected}
\alias{reportExpected}
\title{show the estimated population characteristics from multiple simulated sample}
\usage{
reportExpected(expectedResult = doExpected(100), showType = "Basic")
}
\arguments{
\item{showType}{"Basic", "CILimits", "NHST", "fDR", "FDR:FMR" \cr
\emph{ or one or two of:} \cr
           "r","p","ci1","ci2", "rp","n" \cr 
           "w","wp","wn", ro","po"}
}
\value{
ggplot2 object - and printed
}
\description{
show the estimated population characteristics from multiple simulated sample
}
