% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportInference.R
\name{reportInference}
\alias{reportInference}
\title{report population estimates from a simulated sample}
\usage{
reportInference(
  analysis = braw.res$result,
  analysisType = "Anova",
  showPowerN = TRUE,
  compact = FALSE
)
}
\arguments{
\item{analysisType}{"Model", "Anova"}
}
\value{
ggplot2 object - and printed
}
\description{
report population estimates from a simulated sample
}
\examples{
reportInference(analysis=doAnalysis())
}
