% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLikelihood.R
\name{reportLikelihood}
\alias{reportLikelihood}
\title{report population estimates from a simulated sample}
\usage{
reportLikelihood(possibleResult = braw.res$possibleResult)
}
\arguments{
\item{analysisType}{"Model", "Anova"}
}
\value{
ggplot2 object - and printed
}
\description{
report population estimates from a simulated sample
}
\examples{
reportInference(analysis=doAnalysis())
}
