% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMetaAnalysis.R
\name{reportMetaMultiple}
\alias{reportMetaMultiple}
\title{report a multiple metaAnalysis samples}
\usage{
reportMetaMultiple(metaResult = braw.res$metaMultiple, reportStats = "Medians")
}
\value{
ggplot2 object - and printed
}
\description{
report a multiple metaAnalysis samples
}
\examples{
reportMetaMultiple(metaResult=doMetaMultiple(),reportStats="Medians")
}
