% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMetaAnalysis.R
\name{reportMetaSingle}
\alias{reportMetaSingle}
\title{report a simulated metaAnalysis sample}
\usage{
reportMetaSingle(metaResult = braw.res$metaSingle, reportStats = "Medians")
}
\value{
ggplot2 object - and printed
}
\description{
report a simulated metaAnalysis sample
}
\examples{
reportMetaSingle(metaResult=doMetaAnalysis(),reportStats="Medians")
}
