% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMultiple.R
\name{reportMultiple}
\alias{reportMultiple}
\title{show the estimated population characteristics from multiple simulated sample}
\usage{
reportMultiple(
  multipleResult = braw.res$multiple,
  showType = "Basic",
  whichEffect = "All",
  effectType = "all",
  reportStats = "Medians",
  compact = FALSE
)
}
\arguments{
\item{showType}{"Basic", "CILimits", "NHST", "Hits" \cr
\emph{ or one or two of:} \cr
           "rs","p","ci1","ci2", "rp","n" \cr 
           "ws","wp","nw", ro","po"}
}
\value{
ggplot2 object - and printed
}
\description{
show the estimated population characteristics from multiple simulated sample
}
