% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSystem.R
\name{reportWorld}
\alias{reportWorld}
\title{report the system}
\usage{
reportWorld(hypothesis = braw.def$hypothesis)
}
\value{
ggplot2 object - and printed
}
\description{
report the system
}
\examples{
reportWorld(hypothesis,design)
}
