% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSystem.R
\name{reportWorldDesign}
\alias{reportWorldDesign}
\title{report the system}
\usage{
reportWorldDesign()
}
\value{
ggplot2 object - and printed
}
\description{
report the system
}
\examples{
reportWorldDesign()
}
