% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{setEffect}
\alias{setEffect}
\title{set default effect}
\usage{
setEffect(
  rIV = braw.def$hypothesis$effect$rIV,
  rIV2 = braw.def$hypothesis$effect$rIV2,
  rIVIV2 = braw.def$hypothesis$effect$rIVIV2,
  rIVIV2DV = braw.def$hypothesis$effect$rIVIV2DV,
  rSD = braw.def$hypothesis$effect$rSD,
  rM1 = braw.def$hypothesis$effect$rM1,
  rM2 = braw.def$hypothesis$effect$rM2,
  Heteroscedasticity = braw.def$hypothesis$effect$Heteroscedasticity,
  ResidDistr = braw.def$hypothesis$effect$ResidDistr,
  world = braw.def$hypothesis$effect$world
)
}
\description{
set default effect
}
