% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{setHypothesis}
\alias{setHypothesis}
\title{set default hypothesis}
\usage{
setHypothesis(
  IV = braw.def$hypothesis$IV,
  IV2 = braw.def$hypothesis$IV2,
  DV = braw.def$hypothesis$DV,
  effect = braw.def$hypothesis$effect,
  layout = braw.def$hypothesis$layout
)
}
\description{
set default hypothesis
}
