% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{setWorld}
\alias{setWorld}
\title{set default world}
\usage{
setWorld(
  worldOn = braw.def$hypothesis$effect$world$worldOn,
  populationPDF = braw.def$hypothesis$effect$world$populationPDF,
  populationRZ = braw.def$hypothesis$effect$world$populationRZ,
  populationPDFk = braw.def$hypothesis$effect$world$populationPDFk,
  populationPDFs = braw.def$hypothesis$effect$world$populationPDFs,
  populationPDFmu = braw.def$hypothesis$effect$world$populationPDFmu,
  populationNullp = braw.def$hypothesis$effect$world$populationNullp,
  populationPDFsample = braw.def$hypothesis$effect$world$populationPDFsample,
  populationSamplemn = braw.def$hypothesis$effect$world$populationSamplemn,
  populationSamplesd = braw.def$hypothesis$effect$world$populationSamplesd,
  populationSamplebias = braw.def$hypothesis$effect$world$populationSamplebias,
  sigOnly = braw.def$hypothesis$effect$world$sigOnly,
  worldAbs = braw.def$hypothesis$effect$world$worldAbs
)
}
\description{
set default world
}
