% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showEffect.R
\name{showEffect}
\alias{showEffect}
\title{export}
\usage{
showEffect(
  r,
  moderator = NULL,
  type = 1,
  useCols = c(TRUE, TRUE, TRUE),
  showValue = TRUE,
  plotArea = NULL,
  g = NULL
)
}
\description{
export
}
