% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showExpected.R
\name{showExpected}
\alias{showExpected}
\title{show the estimated population characteristics from multiple simulated sample}
\usage{
showExpected(
  expectedResult = makeExpected(),
  showType = "Basic",
  dimension = "1D",
  orientation = "vert",
  effectType = "direct",
  showTheory = TRUE
)
}
\arguments{
\item{showType}{"Basic", "CILimits", "NHST", "fDR", "FDR:FMR" \cr
\emph{ or one or two of:} \cr
 "r","p","ci1","ci2", "rp","n" \cr
  "w","wp","wn", ro","po"}

\item{dimension}{"1D", "2D"}

\item{orientation}{"vert", "horz"}
}
\value{
ggplot2 object - and printed
}
\description{
show the estimated population characteristics from multiple simulated sample
}
\examples{
showExpected(expectedResult=makeExpected(),
                       showType="Basic",
                       dimension="1D",
                       orientation="vert",
                       effectType="direct",showTheory=TRUE)
}
