% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSystem.R
\name{showFullWorld}
\alias{showFullWorld}
\title{show a world object}
\usage{
showFullWorld(
  hypothesis = braw.def$hypothesis,
  plotArea = c(0, 0, 1, 1),
  fontScale = 1,
  autoShow = FALSE,
  g = NULL
)
}
\value{
ggplot2 object - and printed
}
\description{
show a world object
}
\examples{
showFullWorld(world=makeWorld())
}
