% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSystem.R
\name{showHypothesis}
\alias{showHypothesis}
\title{show a hypothesis}
\usage{
showHypothesis(
  hypothesis = braw.def$hypothesis,
  evidence = braw.def$evidence,
  doWorld = TRUE,
  showValue = TRUE,
  plotArea = NULL,
  autoShow = FALSE,
  g = NULL
)
}
\value{
ggplot2 object - and printed
}
\description{
show a hypothesis
}
\examples{
showHypothesis(hypothesis=makeHypothesis())
}
