% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showInference.R
\name{showInference}
\alias{showInference}
\title{show the estimated population characteristics from a simulated sample}
\usage{
showInference(
  analysis = makeAnalysis(),
  showType = "Basic",
  dimension = "1D",
  orientation = "vert",
  effectType = "direct",
  showTheory = TRUE
)
}
\arguments{
\item{showType}{"Basic", "CILimits", \cr
"NHST","FDR","FMR",
       \emph{ or one or two of:} \cr
"r","p","ci1","ci2", "rp","n"}

\item{dimension}{"1D", "2D"}

\item{orientation}{"vert", "horz"}
}
\value{
ggplot2 object - and printed
}
\description{
show the estimated population characteristics from a simulated sample
}
\examples{
showInference(analysis=makeAnalysis(),
              showType="Basic",
              dimension="1D",
              orientation="vert",
              effectType="direct",
              showTheory=TRUE)
}
