% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMarginals.R
\name{showMarginals}
\alias{showMarginals}
\title{show a simulated sample}
\usage{
showMarginals(result = braw.res$result, style = "piled")
}
\value{
ggplot2 object - and printed
}
\description{
show a simulated sample
}
\examples{
showMarginals(sample=doSample(),style="piled")
}
