% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMetaAnalysis.R
\name{showMetaMultiple}
\alias{showMetaMultiple}
\title{show a multiple meta-analyses}
\usage{
showMetaMultiple(
  metaResult = braw.res$metaMultiple,
  showType = NULL,
  dimension = "1D",
  orientation = "vert"
)
}
\value{
ggplot2 object - and printed
}
\description{
show a multiple meta-analyses
}
\examples{
showMetaMultiple<-function(metaResult=doMetaAnalysis(),showType=NULL,dimension="1D",orientation="vert")
}
