% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMultiple.R
\name{showMultiple}
\alias{showMultiple}
\title{show the estimated population characteristics from multiple simulated sample}
\usage{
showMultiple(
  multipleResult = braw.res$multiple,
  showType = "Basic",
  dimension = "1D",
  orientation = "vert",
  whichEffect = "All",
  effectType = "all",
  showTheory = braw.env$showTheory,
  showData = TRUE,
  showLegend = TRUE
)
}
\arguments{
\item{showType}{"Basic", "p(sig)", "CILimits", "NHST", "Hits", "Misses" \cr
\emph{ or one or two of:} \cr
 "rs","p","ci1","ci2", "rp","n" \cr
  "ws","wp","nw", ro","po"}

\item{dimension}{"1D", "2D"}

\item{orientation}{"vert", "horz"}
}
\value{
ggplot2 object - and printed
}
\description{
show the estimated population characteristics from multiple simulated sample
}
\examples{
showMultiple(multipleResult=doMultiple(),
                       showType="Basic",
                       dimension="1D",
                       orientation="vert",
                       effectType="direct",showTheory=TRUE)
}
