% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPossible.R
\name{showPossible}
\alias{showPossible}
\title{show the likelihood theory}
\usage{
showPossible(
  possibleResult = makePossible(),
  cutaway = FALSE,
  view = "3D",
  axisScale = 1,
  azimuth = 50,
  elevation = 5,
  distance = 2
)
}
\arguments{
\item{view}{"3D","2D"}
}
\value{
ggplot2 object - and printed
}
\description{
show the likelihood theory
}
\examples{
showPossible <- function(possibleResult=makePossible(),
                       cutaway=FALSE,view="3D",axisScale=1,
                       azimuth=50,elevation=5,distance=2)
}
