% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSystem.R
\name{showPrediction}
\alias{showPrediction}
\title{show the prediction corresponding to a hypothesis & design}
\usage{
showPrediction(
  hypothesis = makeHypothesis(),
  design = makeDesign(),
  evidence = makeEvidence()
)
}
\value{
ggplot2 object - and printed
}
\description{
show the prediction corresponding to a hypothesis & design
}
\examples{
showPrediction(hypothesis=makeHypothesis()=makeDesign(),evidence=makeEvidence())
}
