% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showResult.R
\name{showSingle}
\alias{showSingle}
\title{show the analysis of a simulated sample}
\usage{
showSingle(
  result = braw.res$result,
  show = "describe",
  showType = "Basic",
  dimension = "1D"
)
}
\value{
ggplot2 object - and printed
}
\description{
show the analysis of a simulated sample
}
\examples{
showSingle(analysis=doAnalysis(),show="describe",showType="Basic",dimension="1D")
}
