% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSystem.R
\name{showSystem}
\alias{showSystem}
\title{show a system - hypothesis & design}
\usage{
showSystem(
  whichShow = "all",
  hypothesis = braw.def$hypothesis,
  design = braw.def$design,
  evidence = braw.def$evidence
)
}
\value{
ggplot2 object - and printed
}
\description{
show a system - hypothesis & design
}
\examples{
showSystem(whichShow="all",hypothesis=makeHypothesis(),design=makeDesign())
}
