% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showVariable.R
\name{showVariable}
\alias{showVariable}
\title{show a variable object}
\usage{
showVariable(
  variable = makeVariable("test"),
  sample = NULL,
  plotArea = NULL,
  g = NULL
)
}
\arguments{
\item{variable}{a variable object}
}
\value{
a ggplot2 object
}
\description{
show a variable object
}
\examples{
variable<-showVariable(variable=makeVariable())
}
