% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSystem.R
\name{showWorld}
\alias{showWorld}
\title{show a world object}
\usage{
showWorld(world = makeWorld())
}
\value{
ggplot2 object - and printed
}
\description{
show a world object
}
\examples{
showWorld(world=makeWorld())
}
