% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSystem.R
\name{showWorldSampling}
\alias{showWorldSampling}
\title{show the prediction corresponding to a hypothesis & design}
\usage{
showWorldSampling(
  hypothesis = makeHypothesis(),
  design = makeDesign(),
  sigOnly = FALSE
)
}
\value{
ggplot2 object - and printed
}
\description{
show the prediction corresponding to a hypothesis & design
}
\examples{
showWorldSampling(hypothesis=makeHypothesis(),design=makeDesign(),sigOnly=FALSE)
}
