% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicPlot.R
\name{svgBox}
\alias{svgBox}
\title{set up for a meta-analysis}
\usage{
svgBox(height = NULL, aspect = 1.3, fontScale = 1.5)
}
\value{
changes plotting environment
}
\description{
set up for a meta-analysis
}
\examples{
svgB0x(height=200)
}
