% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{makeDesign}
\alias{makeDesign}
\title{make a design}
\usage{
makeDesign(
  sN = 42,
  sMethod = makeSampling("Random"),
  sMethodSeverity = 0.1,
  sNRand = FALSE,
  sNRandSD = 33.3,
  sNRandDist = "Gamma",
  sIV1Use = "Between",
  sIV2Use = "Between",
  sWithinCor = 0.5,
  sBudgetOn = FALSE,
  sNBudget = 1000,
  sIVRangeOn = FALSE,
  sIVRange = c(-1, 1) * 4,
  sIV2RangeOn = FALSE,
  sIV2Range = c(-1, 1) * 4,
  sDVRange = c(-1, 1) * 4,
  sRangeProb = 1,
  sRangeVary = 0,
  sDependence = 0,
  sOutliers = 0,
  sNonResponse = 0,
  sCheating = "None",
  sCheatingAttempts = 10,
  sCheatingLimit = "Fixed",
  sCheatingBudget = 1000,
  sCheatingFixedPop = TRUE,
  Replication = makeReplication(FALSE)
)
}
\arguments{
\item{sMethod}{sampling method object}

\item{sIV1Use}{"Between","Within"}

\item{sCheating}{"None","Grow","Prune","Replace","Retry"}

\item{sCheatingLimit}{"Fixed","Budget"}
}
\value{
a design object
}
\description{
make a design
}
\examples{
makeDesign(sN=42, sMethod=makeSampling("Random"),sMethodSeverity=0.1,
           sNRand=FALSE,sNRandK=2,sNRandDist="Gamma",
           sBudgetOn=FALSE,sNBudget=1000,
           sIV1Use="Between",sIV2Use="Between",  sWithinCor=0.5,
           
           sRangeOn=FALSE, sIVRange=c(-3,3), sDVRange=c(-3,3), sRangeProb=1,
           sDependence=0, sOutliers=0, sNonResponse=0,
           
           sCheating="None", sCheatingAttempts=5,
           sCheatingLimit="Fixed", sCheatingBudget=1000,
           
           Replication=makeReplication()
)
}
\seealso{
[showDesign()]
}
