% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doExplore.R
\name{makeExplore}
\alias{makeExplore}
\title{make multiple samples whilst varying a parameter}
\usage{
makeExplore(
  exploreType = "n",
  exploreNPoints = 11,
  vals = NULL,
  minVal = NA,
  maxVal = NA,
  xlog = FALSE
)
}
\arguments{
\item{exploreType}{"rIV","Heteroscedasticity","rIV2","rIVIV2","rIVIV2DV" \cr
"pNull","Lambda" \cr
"n","Method","Usage","WithinCorr","ClusterRad","SampleSD" \cr
 "Dependence","Outliers","NonResponse","IVRange","IVRangeC","IVRangeE","DVRange" \cr
 "Cheating","CheatingAmount" \cr
 "Alpha","Transform","InteractionOn" \cr
 "Power","Keep","Repeats" \cr}
}
\value{
explore object
}
\description{
make multiple samples whilst varying a parameter
}
\examples{
explore<-makeExplore(exploreType="n",exploreNPoints=11,
                             minVal=10,maxVal=250,xlog=FALSE)
}
\seealso{
doExplore()

showExplore()

reportExplore()
}
