% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{makeHypothesis}
\alias{makeHypothesis}
\title{make a hypothesis object}
\usage{
makeHypothesis(
  IV = makeVariable("IV"),
  IV2 = NULL,
  DV = makeVariable("DV"),
  effect = makeEffect(),
  layout = "normal"
)
}
\value{
a hypothesis object
}
\description{
make a hypothesis object
}
\examples{
makeHypothesis(IV=makeVariable("IV"),IV2=NULL,DV=makeVariable("DV"),effect=makeEffect())
}
\seealso{
showHypothesis()
}
