% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMetaAnalysis.R
\name{makeMetaAnalysis}
\alias{makeMetaAnalysis}
\title{set up for a meta-analysis}
\usage{
makeMetaAnalysis(
  On = FALSE,
  nstudies = 10,
  analysisType = "random",
  analysisVar = "sd",
  method = "MLE",
  analysisPrior = "none",
  modelNulls = FALSE,
  modelPDF = "All",
  sourceBias = FALSE,
  analyseBias = FALSE
)
}
\value{
metaAnalysis object
}
\description{
set up for a meta-analysis
}
\examples{
makeMetaAnalysis<-function(On=TRUE,nstudies=100,
analysisType="random",analysisVar="sd",
method="MLE",analysisPrior="none",
modelNulls=FALSE,modelPDF="All",
sourceBias=FALSE,
analyseBias=FALSE)
}
