% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePossible.R
\name{makePossible}
\alias{makePossible}
\title{likelihood theory}
\usage{
makePossible(
  targetSample = NULL,
  UseSource = "world",
  targetPopulation = NULL,
  UsePrior = "none",
  prior = getWorld("Psych"),
  sims = braw.res$multiple$result,
  sigOnly = FALSE,
  sigOnlyCompensate = FALSE,
  hypothesis = braw.def$hypothesis,
  design = braw.def$design,
  simSlice = 0.1,
  correction = TRUE
)
}
\arguments{
\item{typePossible}{"Samples","Populations"}
}
\value{
possible object
}
\description{
likelihood theory
}
\examples{
makePossible<-function(targetSample=NULL,UseSource="world",
targetPopulation=NULL,UsePrior="none",prior=getWorld("Psych"),
sims=braw.res$multiple$result,sigOnly=FALSE,sigOnlyCompensate=FALSE,
typePossible="Samples",
hypothesis=makeHypothesis(),design=makeDesign(),
simSlice=0.1,correction=TRUE)
}
\seealso{
showPossible()
}
