% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{makeReplication}
\alias{makeReplication}
\title{make a replication object}
\usage{
makeReplication(
  On = TRUE,
  Repeats = 1,
  Keep = "Cautious",
  RepAlpha = 0.05,
  PowerOn = TRUE,
  Power = 0.8,
  Tails = 2,
  PowerPrior = "None",
  UseLikelihood = FALSE,
  replicateAll = FALSE,
  forceSigOriginal = FALSE,
  forceSign = TRUE,
  maxN = 2000,
  BudgetType = "Unlimited",
  Budget = 1000,
  RepNoStudies = 1
)
}
\arguments{
\item{Keep}{"Cautious", "Last", "LargeN", "SmallP", "Median"}

\item{PowerPrior}{"None", "World", "Prior"}

\item{BudgetType}{"Fixed", "Unlimited"}
}
\value{
a replication object
}
\description{
make a replication object
}
\examples{
makeReplication(On=TRUE,Repeats=1,Keep="Cautious",RepAlpha=0.05,
                PowerOn=TRUE,Power=0.8,Tails=2,PowerPrior="None",
                forceSigOriginal=FALSE,forceSign=TRUE,
                BudgetType="Unlimited",Budget=1000
                )
}
