% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVariable.R
\name{makeVariable}
\alias{makeVariable}
\title{make a variable}
\usage{
makeVariable(
  name,
  type = "Interval",
  mu = 0,
  sd = 1,
  skew = 0,
  kurtosis = 0,
  noDigits = NA,
  nlevs = 7,
  iqr = 3,
  median = NULL,
  ordSource = "discrete",
  ordProportions = NA,
  ncats = 2,
  cases = c("C1", "C2"),
  proportions = c(1, 1),
  catSource = "discrete",
  deploy = "Between",
  targetDeploys = "",
  process = "sim"
)
}
\arguments{
\item{name}{- a character string}

\item{type}{"Interval","Ordinal","Categorical"}

\item{ordSource}{"discrete", "continuous"}

\item{catSource}{"discrete", "continuous"}
}
\value{
a Variable object
}
\description{
make a variable
}
\examples{
makeVariable(name,type="Interval",
             mu=0,sd=1,skew=0,kurtosis=0,
             nlevs=7,iqr=3,median=NULL,ordProportions=NA,ordSource="discrete",
             ncats=2,cases=c("C1","C2"),proportions=c(1,1),catSource="discrete"
             )
}
\seealso{
showVariable(variable=makeVariable("test"))

getVariable(name)
}
