% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{makeWorld}
\alias{makeWorld}
\title{make a world
 an object that specifies the distribution of effect sizes}
\usage{
makeWorld(
  On = TRUE,
  PDF = "Uniform",
  RZ = "r",
  PDFk = 0,
  PDFshape = 2,
  PDFspread = 0,
  PDFoffset = 0,
  pRplus = 1,
  PDFsample = FALSE,
  PDFsampleRs = 0,
  PDFsampleN = 0,
  PDFsamplebias = FALSE,
  worldAbs = FALSE
)
}
\arguments{
\item{PDF}{"Single","Double","Uniform","Gauss","Exp","Gamma","GenExp"}

\item{RZ}{"r","z"}
}
\value{
a world object
}
\description{
make a world
 an object that specifies the distribution of effect sizes
}
\examples{
makeWorld<-function(On=FALSE,PDF="Single",RZ="r",
                    PDFk=0.2,pRplus=1,worldAbs=FALSE
)
}
\seealso{
showWorld(world=makeWorld())
}
