% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{makeWorld}
\alias{makeWorld}
\title{make a world
 an object that specifies the distribution of effect sizes}
\usage{
makeWorld(
  On = TRUE,
  PDF = "Uniform",
  RZ = "r",
  PDFk = 0,
  PDFs = 2,
  PDFmu = 0,
  pRPlus = 0,
  PDFsample = FALSE,
  PDFsamplemn = 0,
  PDFsamplesd = 0,
  PDFsamplebias = FALSE,
  sigOnly = FALSE,
  worldAbs = FALSE
)
}
\arguments{
\item{PDF}{"Single","Double","Uniform","Gauss","Exp","Gamma","GenExp"}

\item{RZ}{"r","z"}
}
\value{
a world object
}
\description{
make a world
 an object that specifies the distribution of effect sizes
}
\examples{
makeWorld<-function(On=FALSE,PDF="Single",RZ="r",
                    PDFk=0.2,pRPlus=0,worldAbs=FALSE
)
}
\seealso{
showWorld(world=makeWorld())
}
