% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGLM.R
\name{plotGLM}
\alias{plotGLM}
\title{plot a fitted GLM model}
\usage{
plotGLM(lm, whichR = "Unique")
}
\value{
sample object
}
\description{
plot a fitted GLM model
}
\examples{
plotGLM<-function(DV,IVs,result,whichR)
}
