% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGLM.R
\name{reportGLM}
\alias{reportGLM}
\title{report a fitted GLM model}
\usage{
reportGLM(lm, p_or_r = "r")
}
\value{
sample object
}
\description{
report a fitted GLM model
}
\examples{
reportGLM<-function(lm,p_or_r)
}
