% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSample.R
\name{reportSample}
\alias{reportSample}
\title{report a simulated sample}
\usage{
reportSample(sample = makeSample())
}
\value{
ggplot2 object - and printed
}
\description{
report a simulated sample
}
\examples{
reportSample(sample=makeSample())
}
