% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSEMModel.R
\name{reportSEMModel}
\alias{reportSEMModel}
\title{report a fitted SEM model}
\usage{
reportSEMModel(sem, showType = "CF", showFit = TRUE)
}
\value{
sample object
}
\description{
report a fitted SEM model
}
\examples{
reportSEMModel<-function(sem,showType)
}
