% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{setDesign}
\alias{setDesign}
\title{set default design}
\usage{
setDesign(
  sN = braw.def$design$sN,
  sMethod = braw.def$design$sMethod,
  sMethodSeverity = braw.def$design$sMethodSeverity,
  sNRand = braw.def$design$sNRand,
  sNRandSD = braw.def$design$sNRandSD,
  sNRandDist = braw.def$design$sNRandDist,
  sIV1Use = braw.def$design$sIV1Use,
  sIV2Use = braw.def$design$sIV2Use,
  sWithinCor = braw.def$design$sWithinCor,
  sBudgetOn = braw.def$design$sBudgetOn,
  sNBudget = braw.def$design$sNBudget,
  sIVRangeOn = braw.def$design$sIVRangeOn,
  sIVRange = braw.def$design$sIVRange,
  sIV2RangeOn = braw.def$design$sIV2RangeOn,
  sIV2Range = braw.def$design$sIV2Range,
  sDVRange = braw.def$design$sDVRange,
  sRangeProb = braw.def$design$sRangeProb,
  sRangeVary = braw.def$design$sRangeVary,
  sDependence = braw.def$design$sDependence,
  sOutliers = braw.def$design$sOutliers,
  sNonResponse = braw.def$design$sNonResponse,
  sCheating = braw.def$design$sCheating,
  sCheatingAttempts = braw.def$design$sCheatingAttempts,
  sCheatingLimit = braw.def$design$sCheatingLimit,
  sCheatingBudget = braw.def$design$sCheatingBudget,
  Replication = braw.def$design$Replication
)
}
\description{
set default design
}
