% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{setWorld}
\alias{setWorld}
\title{set default world}
\usage{
setWorld(
  On = braw.def$hypothesis$effect$world$On,
  PDF = braw.def$hypothesis$effect$world$PDF,
  RZ = braw.def$hypothesis$effect$world$RZ,
  PDFk = braw.def$hypothesis$effect$world$PDFk,
  PDFs = braw.def$hypothesis$effect$world$PDFs,
  PDFmu = braw.def$hypothesis$effect$world$PDFmu,
  pRPlus = braw.def$hypothesis$effect$world$pRPlus,
  PDFsample = braw.def$hypothesis$effect$world$PDFsample,
  PDFsamplemn = braw.def$hypothesis$effect$world$PDFsamplemn,
  PDFsamplesd = braw.def$hypothesis$effect$world$PDFsamplesd,
  PDFsamplebias = braw.def$hypothesis$effect$world$PDFsamplebias,
  sigOnly = braw.def$hypothesis$effect$world$sigOnly,
  worldAbs = braw.def$hypothesis$effect$world$worldAbs
)
}
\description{
set default world
}
