% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSystem.R
\name{setWorld}
\alias{setWorld}
\title{set default world}
\usage{
setWorld(
  On = braw.def$hypothesis$effect$world$On,
  PDF = braw.def$hypothesis$effect$world$PDF,
  RZ = braw.def$hypothesis$effect$world$RZ,
  PDFk = braw.def$hypothesis$effect$world$PDFk,
  PDFshape = braw.def$hypothesis$effect$world$PDFshape,
  PDFspread = braw.def$hypothesis$effect$world$PDFspread,
  PDFoffset = braw.def$hypothesis$effect$world$PDFoffset,
  pRplus = braw.def$hypothesis$effect$world$pRplus,
  PDFsample = braw.def$hypothesis$effect$world$PDFsample,
  PDFsampleRs = braw.def$hypothesis$effect$world$PDFsampleRs,
  PDFsampleN = braw.def$hypothesis$effect$world$PDFsampleN,
  PDFsamplebias = braw.def$hypothesis$effect$world$PDFsamplebias,
  worldAbs = braw.def$hypothesis$effect$world$worldAbs
)
}
\description{
set default world
}
