% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDescription.R
\name{showDescription}
\alias{showDescription}
\title{show the sample effect-size analysis of a simulated sample}
\usage{
showDescription(
  analysis = braw.res$result,
  whichEffect = NULL,
  plotArea = c(0, 0, 1, 1),
  dataOnly = FALSE,
  g = NULL
)
}
\value{
ggplot2 object - and printed
}
\description{
show the sample effect-size analysis of a simulated sample
}
\examples{
showDescription(analysis=doAnalysis())
}
