% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showExplore.R
\name{showExplore}
\alias{showExplore}
\title{show the estimated population characteristics from varying parameter}
\usage{
showExplore(
  exploreResult = doExplore(),
  showType = "r",
  showTheory = TRUE,
  effectType = "unique",
  whichEffect = "All"
)
}
\arguments{
\item{showType}{"r","p","w","wn", "p(sig)" \cr
"NHST", "Hits","Misses"}
}
\value{
ggplot2 object - and printed
}
\description{
show the estimated population characteristics from varying parameter
}
\examples{
showExplore(exploreResult=doExplore(),
                       showType="r",
                       effectType="unique",whichEffect="All")
}
