% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showExplore.R
\name{showExplore}
\alias{showExplore}
\title{show the estimated population characteristics from varying parameter}
\usage{
showExplore(
  exploreResult = makeExplore(autoShow = TRUE),
  showType = "r",
  ylog = FALSE,
  whichEffect = "All",
  effectType = "All"
)
}
\arguments{
\item{showType}{"r","p","n","w", "p(sig)" \cr
"NHSTErrors", "FDR","FDR;FMR"}
}
\value{
ggplot2 object - and printed
}
\description{
show the estimated population characteristics from varying parameter
}
\examples{
showExplore(exploreResult=makeExplore(),
                       showType="r",
                       ylog=FALSE,
                       whichEffect="All",effectType="All")
}
