% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showExplore.R
\name{showExplore}
\alias{showExplore}
\title{show the estimated population characteristics from varying parameter}
\usage{
showExplore(
  exploreResult = braw.res$explore,
  showType = "Basic",
  dimension = "1D",
  showTheory = FALSE,
  effectType = "unique",
  whichEffect = "All",
  quantileShow = 0.5,
  autoYlim = TRUE,
  showHist = TRUE,
  fixNulls = TRUE
)
}
\arguments{
\item{showType}{"Basic","p(sig)","NHST", "Hits","Misses" \cr
or one or two of: "rs","p","wp","ws","nw" eg "p;ws"}
}
\value{
ggplot2 object - and printed
}
\description{
show the estimated population characteristics from varying parameter
}
\examples{
showExplore(exploreResult=doExplore(),
                       showType="Basic",dimension="1D",
                       effectType="unique",whichEffect="All",
                       quantileShow=0.5,autoYlim=TRUE,showHist=TRUE)
}
