% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showInference.R
\name{showInference}
\alias{showInference}
\title{show the estimated population characteristics from a simulated sample}
\usage{
showInference(
  analysis = braw.res$result,
  showType = "Basic",
  dimension = "1D",
  orientation = "vert",
  whichEffect = "All",
  effectType = "all",
  showTheory = braw.env$showTheory,
  showData = TRUE,
  showLegend = FALSE,
  sequence = FALSE,
  showYaxis = TRUE
)
}
\arguments{
\item{showType}{"Basic", "CILimits", \cr
"NHST","Hits","Misses",
       \emph{ or one or two of:} \cr
"rs","p","ci1","ci2", "rp","n"}

\item{dimension}{"1D", "2D"}

\item{orientation}{"vert", "horz"}
}
\value{
ggplot2 object - and printed
}
\description{
show the estimated population characteristics from a simulated sample
}
\examples{
showInference(analysis=doAnalysis(),
              showType="Basic",
              dimension="1D",
              orientation="vert",
              whichEffect="Main 1",
              showTheory=TRUE)
}
