% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMetaAnalysis.R
\name{showMetaSingle}
\alias{showMetaSingle}
\title{show a single meta-analysis}
\usage{
showMetaSingle(
  metaResult = braw.res$metaSingle,
  showType = "n",
  showData = TRUE,
  showTheory = TRUE,
  limitNumber = 0,
  xRange = "full",
  fixedYlim = braw.env$fixedYlim,
  fill = NULL,
  alpha = NULL
)
}
\value{
ggplot2 object - and printed
}
\description{
show a single meta-analysis
}
\examples{
showMetaSingle(metaResult=doMetaAnalysis(),showType="n",showTheory=FALSE)
}
