% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSystem.R
\name{showPopulation}
\alias{showPopulation}
\title{show the population corresponding to a hypothesis object}
\usage{
showPopulation(
  hypothesis = braw.def$hypothesis,
  plotArea = c(0, 0, 1, 1),
  autoShow = FALSE,
  g = NULL
)
}
\value{
ggplot2 object - and printed
}
\description{
show the population corresponding to a hypothesis object
}
\examples{
showPopulation(hypothesis=makeHypothesis())
}
