% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPossible.R
\name{showPossible}
\alias{showPossible}
\title{show the likelihood theory}
\usage{
showPossible(
  possibleResult = NULL,
  showType = "Populations",
  cutaway = FALSE,
  walls = "both",
  showP = 0,
  doTextResult = TRUE,
  showSig = FALSE,
  view = "3D",
  axisScale = 1,
  showEvLk = FALSE,
  azimuth = NULL,
  elevation = 15,
  distance = 8
)
}
\arguments{
\item{view}{"3D","2D"}
}
\value{
ggplot2 object - and printed
}
\description{
show the likelihood theory
}
\examples{
showPossible <- function(possibleResult=makePossible(),
                       showType="Populations",
                       cutaway=FALSE,walls=TRUE,showP=0,
                       view="3D",axisScale=1,
                       azimuth=30,elevation=5,distance=2)
}
