% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSystem.R
\name{showWorld}
\alias{showWorld}
\title{show a world object}
\usage{
showWorld(
  hypothesis = braw.def$hypothesis,
  joinNulls = TRUE,
  showSingle = NULL,
  totalArea = 1,
  plotArea = c(0, 0, 1, 1),
  fontScale = 1,
  autoShow = FALSE,
  g = NULL
)
}
\value{
ggplot2 object - and printed
}
\description{
show a world object
}
\examples{
showWorld(world=makeWorld())
}
